Add support for dependency injection frameworks which favor the template method pattern to supply dependencies instead of constructor/method/field injection.
